+++ partiql-lang/src/main/kotlin/org/partiql/lang/eval/physical/PhysicalPlanCompilerImpl.kt
@@ -1833,15 +1833,10 @@ internal class PhysicalPlanCompilerImpl(
                     expr.value.hour.value.toInt(),
                     expr.value.minute.value.toInt(),
                     expr.value.second.decimalValue,
-                    if (expr.value.tzSign != null) {
-                        val totalMinute = expr.value.tzMinutes!!.value + expr.value.tzHour!!.value * 60
-                        if (totalMinute == 0L && expr.value.tzSign!!.text == "-") {
-                            TimeZone.UnknownTimeZone
-                        } else {
-                            TimeZone.UtcOffset.of(totalMinute.toInt())
-                        }
-                    } else {
-                        null
+                    when (val timezone = expr.value.timezone) {
+                        is PartiqlPhysical.Timezone.UnknownTimezone -> TimeZone.UnknownTimeZone
+                        is PartiqlPhysical.Timezone.UtcOffset -> TimeZone.UtcOffset.of(timezone.offsetMinutes.value.toInt())
+                        null -> null
                     },
                     expr.value.precision?.value?.toInt()
                 )