+++ partiql-lang/src/test/kotlin/org/partiql/lang/eval/builtins/TimestampParserTest.kt
@@ -6,9 +6,9 @@ import junitparams.Parameters
 import junitparams.naming.TestCaseName
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.partiql.lang.datetime.TimestampParser
 import org.partiql.lang.errors.ErrorCode
 import org.partiql.lang.eval.EvaluationException
-import org.partiql.lang.eval.builtins.internal.TimestampParser
 import java.lang.reflect.Type
 import java.time.format.DateTimeParseException
 import kotlin.test.assertEquals